[PostInstall] Support multiple pod targets

+++ lib/cocoapods-rome/post_install.rb
@@ -1,23 +1,30 @@
 Pod::HooksManager.register('cocoapods-rome', :post_install) do |installer_context|
-	sandbox_root = installer_context.sandbox_root
-	sandbox = Pod::Sandbox.new(sandbox_root)
+  sandbox_root = installer_context.sandbox_root
+  sandbox = Pod::Sandbox.new(sandbox_root)
 
-	build_dir = Pathname('build')
-	destination = Pathname('Rome')
+  build_dir = Pathname('build')
+  destination = Pathname('Rome')
 
-	Pod::UI.puts 'Building frameworks'
+  Pod::UI.puts 'Building frameworks'
 
-	build_dir.rmtree
-	Dir.chdir(sandbox.project_path.dirname) do
-		Pod::Executable.execute_command 'xcodebuild', %(-project #{sandbox.project_path.basename} -scheme Pods -configuration Release), true
-	end
+  build_dir.rmtree if build_dir.directory?
+  Dir.chdir(sandbox.project_path.dirname) do
+    targets = installer_context.umbrella_targets.select { |t| t.specs.any? }.map(&:cocoapods_target_label)
+    targets.each do |target|
+      Pod::Executable.execute_command 'xcodebuild', %(-project #{sandbox.project_path.basename} -scheme #{target} -configuration Release), true
+    end
+  end
 
-	frameworks = Pathname.glob('build/Release/Pods/*.framework')
+  frameworks = Pathname.glob('build/Release*/Pods*/*.framework')
+  frameworks_by_target = frameworks.group_by { |f| f.to_s =~ %r{^build/Release[^/]*/Pods-?([^/]*)/} && destination + $1 }
 
-	Pod::UI.puts "Built #{frameworks.count} #{'frameworks'.pluralize(frameworks.count)}"
-	Pod::UI.puts 'Copying frameworks to `Rome`'
+  Pod::UI.puts "Built #{frameworks.count} #{'frameworks'.pluralize(frameworks.count)}"
+  Pod::UI.puts "Copying frameworks to `#{destination}`"
 
-	destination.mkpath unless destination.directory?
-	FileUtils.mv(frameworks, 'Rome')
-	build_dir.rmtree
+  destination.rmtree if destination.directory?
+  frameworks_by_target.each do |dest, framework|
+    FileUtils.mkdir_p dest
+    FileUtils.mv framework, dest
+  end
+  build_dir.rmtree
 end
